requirements.txt
apache-airflow-providers-snowflake
dbt-snowflake==1.8.0